backend/requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# ─── Utilities ─────────────────────────────────────────────────────────────────
Pillow>=10.0.0
python-dotenv>=1.0.0

# ─── NLP / Chatbot ─────────────────────────────────────────────────────────────
nltk>=3.8.1

# ─── Data / ML ─────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ─── Dev Tools ─────────────────────────────────────────────────────────────────
# Uncomment below for production:
# gunicorn>=21.2.0
# psycopg2-binary>=2.9.9     # PostgreSQL (replace SQLite in prod)
# whitenoise>=6.6.0          # Static file serving
